+++ src/client/app/components/Button.js
@@ -3,9 +3,10 @@ import React from 'react'
 const Button = props => {
     return (
         <div className={`${props.selector}_div`}>
-            <button className={`${props.selector}_button`}
-            onClick={props.click}>
-            {props.prompt}
+            <button
+                className={`${props.selector}_button`}
+                onClick={props.click} >
+                {props.prompt}
             </button>
         </div>
     )
