+++ src/client/app/components/Input.js
@@ -6,8 +6,16 @@ const Input = props => {
             <p className={`${props.selector}_p`}>
                 {props.prompt}
             </p>
-            <input className={`${props.selector}_input`} type="text" value={props.value} onChange={props.change}/>
-            <button className={`${props.selector}_button`} onClick={props.submit}>Submit</button>
+            <input
+                className={`${props.selector}_input`}
+                type="text"
+                value={props.value}
+                onChange={props.change} />
+            <button
+                className={`${props.selector}_button`}
+                onClick={props.submit} >
+                Submit
+            </button>
         </div>
     )
 }
