+++ src/client/app/containers/DataTableContainer.js
@@ -1,32 +1,33 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import { socket } from '../store/socket'
 import ReactTable from 'react-table'
-import 'react-table/react-table.css'
+
+import { socket } from '../store/socket'
+
 // library
 import { isTrue } from '../services'
+
 // action creators
 import { renderTable } from '../store/reducers/dataTable'
+
 // action keys
 import {
     LOAD_TABLE,
     RESPONSE_RESTAPI
 } from '../store/actions/'
 
-class DataTable extends Component {
-    constructor(props) {
-        super(props)
-    }
+// css stuff down here
+import 'react-table/react-table.css'
 
+class DataTableContainer extends Component {
     render() {
         const parentClass = this.props.visible ?
             "dataTable" : "hidden"
         return (
             <div className={parentClass}>
                 <ReactTable
-                data={this.props.data}
-                columns={this.props.columns}
-                />
+                    data={this.props.data}
+                    columns={this.props.columns} />
             </div>
         )
     }
